# generated/openapi-admin.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/rental-orders": {
      "get": {
        "operationId": "RentalController_list[0]",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "RESERVED",
                "CONFIRMED",
                "ACTIVE",
                "RETURNED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "UNPAID",
                "PARTIALLY_PAID",
                "PAID",
                "REFUNDED",
                "PARTIALLY_REFUNDED"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderPageResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List/search orders; use from/until for calendar overlap queries",
        "tags": [
          "Admin - Rental Orders"
        ]
      },
      "post": {
        "operationId": "RentalController_create[0]",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Recommended for retries from admin/FE.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentalOrderReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create order + price snapshots + physical allocations atomically",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}": {
      "get": {
        "operationId": "RentalController_get[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/schedule": {
      "patch": {
        "operationId": "RentalController_reschedule[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRentalReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reschedule within the policy window from original booking; preserve duration and recheck overlap",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/confirm": {
      "post": {
        "operationId": "RentalController_confirm[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRentalReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid collateral, evidence, or order state (including repeated confirmation)."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Requires rentals.confirm."
          },
          "404": {
            "description": "Order does not exist in authenticated shop."
          },
          "413": {
            "description": "Evidence exceeds 15 MiB."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Manually acknowledge receipt and collateral; atomically confirm without Payment API. Requires rentals.confirm.",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/confirmation-options": {
      "get": {
        "operationId": "RentalController_confirmationOptions[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationOptionsResDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Requires rentals.confirm."
          },
          "404": {
            "description": "Order does not exist in authenticated shop."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/confirmation/evidence": {
      "get": {
        "operationId": "RentalController_evidence[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evidence image; requires rentals.confirm. Private, no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Requires rentals.confirm."
          },
          "404": {
            "description": "Order or evidence does not exist in authenticated shop."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/start": {
      "post": {
        "operationId": "RentalController_start[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRentalReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/collateral/return": {
      "post": {
        "operationId": "RentalController_returnCollateral[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/return-preview": {
      "get": {
        "operationId": "RentalController_returnPreview[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnedAt",
            "required": false,
            "in": "query",
            "description": "ISO-8601 string of actual return timestamp",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnPreviewResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Preview return details, calculate late fees and list items for inspection",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/return": {
      "post": {
        "operationId": "RentalController_receiveReturn[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnRentalOrderReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Receive rental return with per-item inspection and optional manual charges",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/settle": {
      "post": {
        "operationId": "RentalController_settle[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SettleRentalOrderReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Settle returned rental order; offsets charges against cash deposit, releases document collateral if requested, and completes order",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/settlement/evidence": {
      "get": {
        "operationId": "RentalController_settlementEvidence[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settlement evidence image; requires rentals.settle. Private, no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/cancel": {
      "post": {
        "operationId": "RentalController_cancel[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRentalReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cancel a RESERVED order; confirmed and active orders cannot be cancelled",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
    "/admin/rental-orders/{id}/charges": {
      "post": {
        "operationId": "RentalController_addCharge[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRentalChargeReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kitty Shop API - Admin API",
    "description": "Rental shop admin API for staff & operations (kitty-admin-fe). Tenant scope comes from the authenticated shop membership. Monetary values are decimal-safe values.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API prefix"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Configured application URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RentalOrderPageResDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalOrderListItemResDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaResDto"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "CreateRentalOrderReqDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "rentalStartAt": {
            "type": "string",
            "example": "2026-09-12T03:00:00.000Z"
          },
          "rentalEndAt": {
            "type": "string",
            "example": "2026-09-14T03:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateRentalItemReqDto"
            }
          },
          "charges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalChargeReqDto"
            }
          },
          "discountTotal": {
            "type": "number",
            "default": 0
          },
          "note": {
            "type": "string"
          },
          "internalNote": {
            "type": "string"
          },
          "delivery": {
            "$ref": "#/components/schemas/RentalDeliveryReqDto"
          },
          "collateral": {
            "$ref": "#/components/schemas/RentalCollateralReqDto"
          }
        },
        "required": [
          "customerId",
          "rentalStartAt",
          "rentalEndAt",
          "items"
        ]
      },
...
      "RescheduleRentalReqDto": {
        "type": "object",
        "properties": {
          "rentalStartAt": {
            "type": "string"
          },
          "rentalEndAt": {
            "type": "string"
          }
        },
        "required": [
          "rentalStartAt",
          "rentalEndAt"
        ]
      },
      "ConfirmRentalReqDto": {
        "type": "object",
        "properties": {
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "BANK_TRANSFER"
            ]
          },
          "collateralMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "DOCUMENT"
            ]
          },
          "documentType": {
            "type": "string",
            "enum": [
              "CCCD",
              "GPLX"
            ]
          },
          "collateralAmount": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000,
            "description": "Actual deposit received; may be below the suggested deposit, including zero."
          },
          "note": {
            "type": "string",
            "maxLength": 2000
          },
          "evidence": {
            "type": "string",
            "format": "binary",
            "description": "Optional evidence image, at most 15 MiB."
          }
        },
        "required": [
          "collateralMethod"
        ]
      },
      "ConfirmationOptionsResDto": {
        "type": "object",
        "properties": {
          "allowedMethods": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CASH",
                "DOCUMENT"
              ]
            }
          },
          "allowedDocumentTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CCCD",
                "GPLX"
              ]
            }
          },
          "expectedDeposit": {
            "type": "string"
          },
          "maxEvidenceBytes": {
            "type": "number"
          },
          "evidenceMimeTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "allowedMethods",
          "allowedDocumentTypes",
          "expectedDeposit",
          "maxEvidenceBytes",
          "evidenceMimeTypes"
        ]
      },
      "TransitionRentalReqDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "ReturnPreviewResDto": {
        "type": "object",
        "properties": {
          "rentalEndAt": {
            "type": "string",
            "format": "date-time"
          },
          "actualReturnedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lateDays": {
            "type": "number"
          },
          "dailyLateFeePerSet": {
            "type": "number"
          },
          "lateFee": {
            "type": "string",
            "example": "20000.00"
          },
          "additionalRentalFee": {
            "type": "string",
            "example": "0.00"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnPreviewItemDto"
            }
          }
        },
        "required": [
          "rentalEndAt",
          "actualReturnedAt",
          "lateDays",
          "dailyLateFeePerSet",
          "lateFee",
          "additionalRentalFee",
          "items"
        ]
      },
      "ReturnRentalOrderReqDto": {
        "type": "object",
        "properties": {
          "actualReturnedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Thời điểm trả thực tế (mặc định là hiện tại)"
          },
          "inspections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnInspectionItemReqDto"
            }
          },
          "manualCharges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalChargeReqDto"
            }
          },
          "note": {
            "type": "string",
            "description": "Ghi chú trả hàng"
          }
        },
        "required": [
          "inspections"
        ]
      },
...
    }
  }
}